requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.5
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
plotly==5.18.0
pandas==2.1.4
python-dateutil==2.8.2
requests==2.31.0
python-jose==3.3.0
Flask-Migrate==4.0.5
